[Feat] MapSection 분리
