+++ src/entities/recovery/incidents.ts
@@ -1,31 +1,3 @@
-import { getResponse } from "@shared/lib/instance";
-
-export interface IncidentItem {
-  id: number;
-  occurred_at: string;
-  address: string;
-  lat: number;
-  lng: number;
-  cause: string;
-  method: string;
-  status: "UNDER_REPAIR" | "TEMP_REPAIRED" | "RECOVERED";
-  images_count: number;
-  distance_m: number;
-}
-
-export interface IncidentsResponse {
-  status: string;
-  message: string;
-  code: number;
-  data: {
-    items: IncidentItem[];
-    count: number;
-  };
-}
-
-// 사고 목록 조회 API
-export const getIncidents = async (statuses: ("UNDER_REPAIR" | "TEMP_REPAIRED" | "RECOVERED")[] = ["UNDER_REPAIR", "TEMP_REPAIRED"]): Promise<IncidentsResponse | null> => {
-  const statusParam = statuses.join(',');
-  const url = `/api/v1/incidents?status=${statusParam}`;
-  return await getResponse<IncidentsResponse>(url);
-};
+// Re-export from API for backward compatibility
+export { getIncidents } from "./api";
+export type { IncidentItem, IncidentsResponse } from "./response";
