+++ src/entities/recovery/model.ts
@@ -1,0 +1,11 @@
+// Recovery API Request Parameter Types
+
+// getIncidents API parameters
+export interface GetIncidentsParams {
+  statuses?: ("UNDER_REPAIR" | "TEMP_REPAIRED" | "RECOVERED")[];
+}
+
+// Recovery status types
+export type RecoveryStatus = "UNDER_REPAIR" | "TEMP_REPAIRED" | "RECOVERED";
+
+// No parameters needed for getRecoveryStatus API
