+++ src/entities/sinkhole/api.ts
@@ -1,49 +1,45 @@
-import { getResponse } from "@shared/lib/instance";
-import type { SelectGradeResponse, Grade } from "./selectgrade";
-import type { SafezoneResponse } from "./safezones";
-import type { DistrictSearchResponse, DistrictSearchParams, SafezoneDistrictsResponse } from "@features/sinkhole-map/constants";
+// Sinkhole Entity API - delegates to feature APIs
+import * as SinkholeMapAPI from "@features/sinkhole-map/api";
+import type { Grade, DistrictSearchParams } from "./model";
+import type { 
+  SelectGradeResponse, 
+  SafezoneResponse, 
+  DistrictSearchResponse, 
+  SafezoneDistrictsResponse 
+} from "./response";
+
+// Re-export types for backward compatibility
+export type { SelectGradeResponse, SafezoneResponse, DistrictSearchResponse, SafezoneDistrictsResponse } from "./response";
+export type { SelectGradeItem, SafezoneItem, DistrictSearchItem, SafezoneDistrictItem } from "./response";
+export type { Grade } from "./model";
 
 // 구 등급 필터 조회 API
 export const getDistrictsGuByGrade = async (grade: Grade): Promise<SelectGradeResponse | null> => {
-  const url = `/api/v1/districts/gu/metrics/by-grade?grade=${grade}`;
-  return await getResponse<SelectGradeResponse>(url);
+  return await SinkholeMapAPI.getDistrictsGuByGrade({ grade });
 };
 
 // 동 등급 필터 조회 API
 export const getDistrictsByGrade = async (grade: Grade): Promise<SelectGradeResponse | null> => {
-  const url = `/api/v1/districts/by-grade?grade=${grade}`;
-  return await getResponse<SelectGradeResponse>(url);
+  return await SinkholeMapAPI.getDistrictsByGrade({ grade });
 };
 
 // 안심존 구 조회 API
 export const getSafezones = async (): Promise<SafezoneResponse | null> => {
-  const url = `/api/v1/districts/safezones/gu`;
-  return await getResponse<SafezoneResponse>(url);
+  return await SinkholeMapAPI.getSafezones();
 };
 
-// 행정동 검색 API
-export const searchDistricts = async (params: DistrictSearchParams): Promise<DistrictSearchResponse | null> => {
-  const searchParams = new URLSearchParams();
-  searchParams.append('q', params.q);
-  if (params.sido) searchParams.append('sido', params.sido);
-  if (params.sigungu) searchParams.append('sigungu', params.sigungu);
-  if (params.limit) searchParams.append('limit', params.limit.toString());
-  
-  const url = `/api/v1/districts/search?${searchParams.toString()}`;
-  return await getResponse<DistrictSearchResponse>(url);
-};
-
-// export const districtSafeGrade = async ()
-
-// G1, G2 등급 안심구역 조회 API (구별)
+// G1, G2 등급 안심구역 조회 API (구별) - getSafezones와 동일
 export const getSafezoneGu = async (): Promise<SafezoneResponse | null> => {
-  const url = `/api/v1/districts/safezones/gu`;
-  return await getResponse<SafezoneResponse>(url);
+  return await SinkholeMapAPI.getSafezoneGu();
 };
 
 // 안심존 행정동 조회 API (동별)
 export const getSafezoneDistricts = async (): Promise<SafezoneDistrictsResponse | null> => {
-  const url = `/api/v1/districts/safezones`;
-  return await getResponse<SafezoneDistrictsResponse>(url);
+  return await SinkholeMapAPI.getSafezoneDistricts();
+};
+
+// 행정동 검색 API
+export const searchDistricts = async (params: DistrictSearchParams): Promise<DistrictSearchResponse | null> => {
+  return await SinkholeMapAPI.searchDistricts(params);
 };
