+++ src/pages/SinkholePage.tsx
@@ -3,9 +3,9 @@ import * as S from "@widget/MapWithOverlay/MapWithOverlay.styles";
 import { PageHeader, SearchBar } from "@shared/components";
 import {
   MainElement,
-  MapSection,
   BackgroundCircles,
 } from "@features/recovery-zone";
+import { SinkholeMapSection } from "@features/sinkhole-map/containers/SinkholeMapSection";
 import { BottomSheet } from "@features/sinkhole-map/containers/BottomSheet";
 import { SelectGradeProvider, useSelectGrade } from "@entities/sinkhole/context";
 import { searchDistricts } from "@entities/sinkhole/api";
@@ -54,9 +54,8 @@ const SinkholePageContent = () => {
         {/* 메인 콘텐츠 컨테이너 (현재 빈 상태) */}
         <MainElement.MainContent />
         {/* 서울 지역 싱크홀 위험도 지도 - 구별 색상 표시 */}
-        <MapSection 
+        <SinkholeMapSection 
           id="main-map"
-          colorMode="risk" 
           forceViewMode={isBadgeActive ? "safezone" : undefined}
         />
       </MainElement.MainWrapper>
